writing line by line as data in string is huge

+++ hardware/stepperCalculator/Commands/StepData.cs
@@ -2,7 +2,7 @@
 {
     public class StepData
     {
-        public int TimeStamp { get; set; }
+        public double TimeStamp { get; set; }
         public double PositionAfterStep { get; set; }
         public double StepTime { get; set; }
         public double SpeedAfterMove { get; set; }
